[PAGOPA-346] fix: Resolved bug on API calls towards GPD service
During the evaluation of the performance test, seemed that the calls of the API related to the payment options ended unsuccessfully due to the fact that no API subscription key was sent. This problem was resolved setting the 'configuration' field on FeignClient annotation for GpdClient to AuthFeignConfig instead of FeignConfig. This class was set because it can also add the header related to the subscription key. The same thing was define for GpsClient, despite the fact that no test was carried out on this class.