[PAGOPA-355] feat: improved performance-tests
This commits includes all the changes made to performance-test folder and the files included in here in order to migrate towards the structure defined in the template project. Also, the configuration files for DEV and UAT tests are rewritten (and so the logic for the retrieving in the scripts) in order to get a tidier command.